+++ lib/src/clojure_lsp/feature/file_management.clj
@@ -262,6 +262,10 @@
       (shared/dissoc-in [:analysis filename])
       (shared/dissoc-in [:findings filename])))
 
+(defn ^:private file-deleted [db* uri filename]
+  (swap! db* db-without-file uri filename)
+  (f.diagnostic/publish-empty-diagnostics! uri @db*))
+
 (defn did-change-watched-files [changes db*]
   (doseq [{:keys [uri type]} changes]
     (case type
@@ -275,16 +279,14 @@
                                db*)))
       :deleted (shared/logging-task
                  :delete-watched-file
-                 (let [filename (shared/uri->filename uri)]
-                   (swap! db* db-without-file uri filename))))))
+                 (file-deleted db* uri (shared/uri->filename uri))))))
 
 (defn did-close [uri db*]
   (let [filename (shared/uri->filename uri)
         source-paths (settings/get @db* [:source-paths])]
     (when (and (not (shared/external-filename? filename source-paths))
                (not (shared/file-exists? (io/file filename))))
-      (swap! db* db-without-file uri filename)
-      (f.diagnostic/publish-empty-diagnostics! uri @db*))))
+      (file-deleted db* uri filename))))
 
 (defn force-get-document-text
   "Get document text from db, if document not found, tries to open the document"
@@ -302,7 +304,9 @@
        (keep (fn [{:keys [old-uri new-uri]}]
                (let [old-filename (shared/uri->filename old-uri)
                      new-ns (shared/uri->namespace new-uri db)
-                     ns-definition (q/find-namespace-definition-by-filename db old-filename)]
-                 (when (and new-ns ns-definition)
-                   (f.rename/rename-element old-uri new-ns db ns-definition :rename-file)))))
+                     old-ns-definition (q/find-namespace-definition-by-filename db old-filename)]
+                 (when (and new-ns
+                            old-ns-definition
+                            (not= new-ns (name (:name old-ns-definition))))
+                   (f.rename/rename-element old-uri new-ns db old-ns-definition :rename-file)))))
        (reduce #(shared/deep-merge %1 %2) {:document-changes []})))