+++ lib/src/clojure_lsp/handlers.clj
@@ -42,22 +42,24 @@
        (reductions +)
        (take 15)))
 
-(defmacro process-after-changes [task-id uri db* & body]
+(defmacro process-after-all-changes [task-id uris db* & body]
   (let [waiting-start-sym (gensym "waiting-start-time")
         start-sym (gensym "start-time")
         backoff-sym (gensym "backoff")
+        uris-sym (gensym "uris")
         process-msg (str task-id " %s")
         wait-and-process-msg (str task-id " %s - waited %s")]
     `(let [~waiting-start-sym (System/nanoTime)]
-       (loop [~backoff-sym backoff-start]
+       (loop [~backoff-sym backoff-start
+              ~uris-sym ~uris]
          (if (> (quot (- (System/nanoTime) ~waiting-start-sym) 1000000) 60000) ; one minute timeout
            ~(with-meta
-              `(logger/warn (format "Timeout in %s waiting for changes to %s" ~task-id ~uri))
+              `(logger/warn (format "Timeout in %s waiting for changes to %s" ~task-id (first ~uris-sym)))
               (meta &form))
-           (if (contains? (:processing-changes @~db*) ~uri)
+           (if-let [processing-uris# (seq (filter (:processing-changes @~db*) ~uris-sym))]
              (do
                (Thread/sleep ~backoff-sym)
-               (recur (min backoff-max (* backoff-mult ~backoff-sym))))
+               (recur (min backoff-max (* backoff-mult ~backoff-sym)) processing-uris#))
              (let [~start-sym (System/nanoTime)
                    result# (do ~@body)]
                ~(with-meta
@@ -69,6 +71,9 @@
                                (shared/format-time-delta-ms ~waiting-start-sym ~start-sym))))
                   (meta &form))
                result#)))))))
+
+(defmacro process-after-changes [task-id uri db* & body]
+  `(process-after-all-changes ~task-id [~uri] ~db* ~@body))
 
 (defn initialize
   [{:keys [db* producer] :as components}
@@ -470,6 +475,7 @@
       (f.linked-editing-range/ranges (:uri text-document) row col db))))
 
 (defn will-rename-files [{:keys [db*]} {:keys [files]}]
-  (shared/logging-task
+  (process-after-all-changes
     :will-rename-files
+    (map :oldUri files) db*
     (f.file-management/will-rename-files files @db*)))