+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   - Fix to update unused-public-var lint on registered keywords as usages change in other files. #1018
   - Fix to navigate to var defined by declare, when there aren't any later defs. #1107
   - Fix to always go to the definition of the correct var imported by potemkin. #1020
+  - Fix to correctly rename namespaces. #1121
 
 ## 2022.06.29-19.32.13